+++ ExpenseTracking.Api/Impl/Cqrs/Employee.cs
@@ -5,7 +5,7 @@ using MediatR;
 namespace ExpenseTracking.Api.Impl.Cqrs;
 
 public record GetAllEmployeesQuery : IRequest<ApiResponse<List<EmployeeResponse>>>;
-public record GetEmployeeByIdQuery(int Id) : IRequest<ApiResponse<EmployeeDetailResponse>>;
+public record GetEmployeeByIdQuery(long Id) : IRequest<ApiResponse<EmployeeDetailResponse>>;
 public record GetEmployeesByParametersQuery(long? DepartmentId, int? MinSalary, int? MaxSalary) : IRequest<ApiResponse<List<EmployeeResponse>>>;
 
 public record CreateEmployeeCommand(
@@ -13,5 +13,5 @@ public record CreateEmployeeCommand(
     EmployeeRequest Employee
 ) : IRequest<ApiResponse<CreateEmployeeResponse>>;
 
-public record UpdateEmployeeCommand(int Id, UpdateEmployeeRequest Employee) : IRequest<ApiResponse>;
-public record DeleteEmployeeCommand(int Id) : IRequest<ApiResponse>;
+public record UpdateEmployeeCommand(long Id, UpdateEmployeeRequest Employee) : IRequest<ApiResponse>;
+public record DeleteEmployeeCommand(long Id) : IRequest<ApiResponse>;
