+++ ExpenseTracking.Api/Impl/Cqrs/ExpenseFile.cs
@@ -5,7 +5,7 @@ using MediatR;
 namespace ExpenseTracking.Api.Impl.Cqrs;
 
 public record GetAllExpenseFilesQuery : IRequest<ApiResponse<List<ExpenseFileResponse>>>;
-public record GetExpenseFileByIdQuery(int Id) : IRequest<ApiResponse<ExpenseFileResponse>>;
+public record GetExpenseFileByIdQuery(long Id) : IRequest<ApiResponse<ExpenseFileResponse>>;
 public record CreateExpenseFileCommand(ExpenseFileRequest ExpenseFile) : IRequest<ApiResponse>;
-public record UpdateExpenseFileCommand(int Id, UpdateExpenseFileRequest ExpenseFile) : IRequest<ApiResponse>;
-public record DeleteExpenseFileCommand(int Id) : IRequest<ApiResponse>;
+public record UpdateExpenseFileCommand(long Id, UpdateExpenseFileRequest ExpenseFile) : IRequest<ApiResponse>;
+public record DeleteExpenseFileCommand(long Id) : IRequest<ApiResponse>;
